(WIP) Fix issues caused by the "edit" method
